s1-dev over

+++ src/com/pplive/liveplatform/util/TimeUtil.java
@@ -3,12 +3,12 @@ package com.pplive.liveplatform.util;
 import java.util.Locale;
 
 public class TimeUtil {
-    public static String stringForTime(int timeMs) {
-        int totalSeconds = timeMs / 1000;
+    public static String stringForTime(long timeMs) {
+        long totalSeconds = timeMs / 1000;
 
-        int seconds = totalSeconds % 60;
-        int minutes = (totalSeconds / 60) % 60;
-        int hours = totalSeconds / 3600;
+        long seconds = totalSeconds % 60;
+        long minutes = (totalSeconds / 60) % 60;
+        long hours = totalSeconds / 3600;
 
         return String.format(Locale.US, "%02d:%02d:%02d", hours, minutes, seconds);
     }
